+++ CHANGELOG.md
@@ -5,6 +5,7 @@ It is based on Tidepool Blip 1.27.
 ## 3.0.4-beta - 2023-02-10
 ### Added
 - YLP-1503 Logout idle user after 30 mn of inactivity
+- YLP-2140 As a HCP, I can only edit my medical records
 
 ### Engineering use
 - YLP-2044 Remove redux settings store
@@ -29,7 +30,12 @@ It is based on Tidepool Blip 1.27.
 - YLP-2121 Removing deprecated @mui/styles library
 - YLP-2139 Remove weekly reports
 - YLP-2143 Migrate TrendsContainer to dumb
+- YLP-2144 Migrate dashboard from blip to Yourloops
 - YLP-2159 Get rid of connectWithTransitionGroup in viz
+- YLP-2162 Migrate CbgDateTrace store from redux to TrendsProvider
+
+### Fixed
+- YLP-2145 Patient cannot see monitoring widgets when he's in a monitoring team and another non-monitoring team
 
 ## 3.0.3 - 2023-02-09
 ### Fixed